Add files via upload
This project involves the comprehensive electrical wiring and system setup for a 10-story building with a total area exceeding 1600 square feet. It was designed in AutoCAD. It includes detailed planning and layout of the building’s structural framework to accommodate electrical installations, strategic placement of electrical fittings and fixtures, and design of conduit pathways for efficient wiring distribution. The project covers wiring for both light and heavy electrical loads, establishing connections between switchboards, Sub-Distribution Boards (SDB), and Main Distribution Boards (MDB), and installation of antenna and telephone systems. Detailed calculations are performed for switchboards, SDB, and MDB to ensure correct sizing and capacity, and appropriate breakers are selected for different groups. The project also includes creating detailed diagrams for distribution boards and the main distribution board, calculating and designing the Single Line Diagram (SLD), and implementing a lightning protection system. An effective earthing system is designed and installed to ensure safety and compliance with regulations. Lighting requirements for living rooms are calculated to ensure adequate illumination. The project involves calculating the photovoltaic (PV) requirements for the building, including battery sizing, and determining the appropriate sizes for solar charge controllers and PV inverters. The rooftop PV system is designed and set up to maximize solar energy capture, and models of the solar PV system and fire safety system are developed to visualize and plan the installation. This project aims to deliver a robust, efficient, and safe electrical system for the building, incorporating modern technologies and best practices in electrical engineering. The zip file which contains the AutoCAD files are locked. If you need this then you can email me.